requirements.txt
requests
python-dotenv
toml
beautifulsoup4
git+https://github.com/deepset-ai/haystack.git@v2.0.0-beta.5
huggingface_hub
sentence-transformers>=2.2.0
